+++ Classes/Command/DeleteExpiredCommand.php
@@ -22,13 +22,15 @@ use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Style\SymfonyStyle;
+use TYPO3\CMS\Core\Resource\FileRepository;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
-use Wacon\Filetransfer\Domain\Repository\UploadRepository;
+use Wacon\Filetransfer\Domain\QueryBuilder\UploadQueryBuilder;
 
 class DeleteExpiredCommand extends Command
 {
     public function __construct(
-        private readonly UploadRepository $uploadRepository
+        private readonly UploadQueryBuilder $uploadQueryBuilder,
+        private readonly FileRepository $fileRepository
     ) {
         parent::__construct();
     }
@@ -57,27 +59,21 @@ class DeleteExpiredCommand extends Command
     {
         $amountOfFiles = 0;
 
-        $this->initRepositories(GeneralUtility::intExplode(',', $input->getArgument('pids'), true));
-        $uploads = $this->uploadRepository->findByAllExpired();
-        $amountOfFiles = count($uploads);
+        $uploads = $this->uploadQueryBuilder->findByAllExpired();
+        $amountOfFiles = $uploads->rowCount();
+        $uploads = $uploads->fetchAllAssociative();
 
         foreach ($uploads as $upload) {
-            /**
-             * @var Upload $upload
-             */
             // Delete file
-            $asset = $upload->getAsset();
+            $asset = $this->getAsset($upload);
 
             if ($asset) {
-                $asset->getOriginalResource()->getOriginalFile()->delete();
+                $asset->getOriginalFile()->delete();
+                $asset->delete();
             }
 
             // Delete record
-            $this->uploadRepository->remove($upload);
-        }
-
-        if ($amountOfFiles > 0) {
-            $this->uploadRepository->commit();
+            $this->uploadQueryBuilder->remove($upload);
         }
 
         // Render answer
@@ -88,13 +84,14 @@ class DeleteExpiredCommand extends Command
     }
 
     /**
-     * Init repositories
-     * @param array $pids
+     * Get asset
+     * @param array $upload
+     * @return \TYPO3\CMS\Core\Resource\FileReference|bool
      */
-    private function initRepositories(array $pids)
+    private function getAsset(array $upload)
     {
-        $querySettings = $this->uploadRepository->createQuery()->getQuerySettings();
-        $querySettings->setStoragePageIds($pids);
-        $this->uploadRepository->setDefaultQuerySettings($querySettings);
+        $assets = $this->fileRepository->findByRelation(UploadQueryBuilder::DB_TABLE, 'asset', $upload['uid']);
+
+        return is_array($assets) ? current($assets) : $assets;
     }
 }